move options error to central location
