+++ server/utils/errors.py
@@ -1,0 +1,7 @@
+class OptionsError(Exception):
+    """
+    Raised when user specified cli options specified fail parsing
+    """
+
+    def __init__(self, message):
+        self.message = message
